Revert "Clean up include_class deprecation notice"
This reverts commit 8d951da86be5e61e17fec4491cb2990b8efb590b.

The RSpec-core commit c07c4cf9b47b9f58b222b77ca13daea11b3d7585
(included in 2.99.0) has changed the deprecation api:
    - def deprecate(deprecated, replacement_or_hash={}, ignore_version=nil)
    - data = Hash === replacement_or_hash
                        ? replacement_or_hash
                        : { :replacement => replacement_or_hash }
    + def deprecate(deprecated, data = {})

The commit that has been reverted changed the second parameter from
    replacement: foo
to
    foo
which is incompatible with the new form, so it has to be reverted.

Now it works with 2.X again.

The tests have also been adjusted accordingly